Skip testing fp16/fp64 types if not supported by a device
The limitation is that wrapper-like types are not properly handled;
the most important one is zip_iterator

+++ test/parallel_api/algorithm/alg.modifying.operations/alg.reverse/reverse.pass.cpp
@@ -101,7 +101,7 @@ main()
     test<std::int32_t>();
     test<std::uint16_t>();
     test<float64_t>();
-    test<wrapper<float64_t>>();
+    test<wrapper<float32_t>>();
 
     return done();
 }
